pyproject.toml
[project]
name = "better-social-notifications"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14, <3.15"
dependencies = [
    "apprise==1.9.7",
    "google-api-python-client==2.188.0",
    "python-dotenv==1.2.1",
    "sqlalchemy==2.0.46",
]
license-files = ["LICENSE"]
authors = [
    { name = "Jack Stockley", email = "jack@jstockley.com"}
]
keywords = ["starter", "template", "python"]
classifiers = [
    "Programming Language :: Python :: 3",
]
[project.urls]
Homepage = "https://github.com/jnstockley/python-starter"
Repository = "https://github.com/jnstockley/python-starter.git"
Issues = "https://github.com/jnstockley/python-starter/issues"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.13.2",
]
